Extract useRouter function
