+++ app.json
@@ -15,9 +15,7 @@
   "formation": {
   },
   "addons": [
-
   ],
   "buildpacks": [
-
   ]
 }
